+++ code/libft/str/ft_strrchr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strrchr.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jmilhas <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: aartiges <aartiges@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/03 18:12:09 by jmilhas           #+#    #+#             */
-/*   Updated: 2021/11/03 18:25:22 by jmilhas          ###   ########.fr       */
+/*   Updated: 2022/02/04 20:23:43 by aartiges         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *str, int c)
 {
-	int	n;
+	size_t	n;
 
 	n = ft_strlen((char *)str);
 	while (n >= 0)
